renaming CTPPSPixelLocal to PixelLocal
