+++ RecoCTPPS/Configuration/python/recoCTPPS_cff.py
@@ -3,7 +3,7 @@ import FWCore.ParameterSet.Config as cms
 from RecoCTPPS.TotemRPLocal.totemRPLocalReconstruction_cff import *
 from RecoCTPPS.TotemRPLocal.ctppsDiamondLocalReconstruction_cff import *
 from RecoCTPPS.TotemRPLocal.ctppsLocalTrackLiteProducer_cfi import ctppsLocalTrackLiteProducer
-from RecoCTPPS.CTPPSPixelLocal.ctppsPixelLocalReconstruction_cff import *
+from RecoCTPPS.PixelLocal.ctppsPixelLocalReconstruction_cff import *
 
 recoCTPPS = cms.Sequence(
     totemRPLocalReconstruction *
